requirements.txt
apache-airflow==2.8.0
apache-airflow-providers-amazon==8.13.0
requests==2.31.0
pandas==2.1.4
pyarrow==14.0.2
boto3==1.34.34
